fixed document deletion bug and rerouting after document deleted

+++ src/components/DocumentItem.vue
@@ -3,7 +3,9 @@
     <button type="button" class="doc-button" @click="goTo">
       {{ document.title }}
     </button>
-    <i @click="deleteDocument" class="material-icons folder-more-icon"> delete </i>
+    <i @click="deleteDocument" class="material-icons folder-more-icon">
+      delete
+    </i>
   </div>
 </template>
 
@@ -44,6 +46,9 @@ export default {
         const docSnap = await getDoc(docRef);
 
         if (docSnap.exists()) {
+          if(this.$route.params.doc == this.docId){
+            this.$router.push("/editor");
+          }
           await this.deleteDocImages(docSnap);
           await deleteDoc(
             doc(db, "users", auth.currentUser.uid, "notes", this.docId)
@@ -53,22 +58,25 @@ export default {
         }
       }
     },
-    deleteDocImages: async function(document){
-      const baseUrl = "https://firebasestorage.googleapis.com/v0/b/darkdown-44b5e.appspot.com/o/";
+    deleteDocImages: async function (document) {
+      const baseUrl =
+        "https://firebasestorage.googleapis.com/v0/b/darkdown-44b5e.appspot.com/o/";
       let removedImagesSrc = document.data().imageurls;
-      removedImagesSrc.forEach(imageUrl => {
-          if (imageUrl.includes("firebasestorage")){
-            console.log("Deleting:", imageUrl)
-            let imagePath = imageUrl.replace(baseUrl,"");
+      if (removedImagesSrc) {
+        removedImagesSrc.forEach((imageUrl) => {
+          if (imageUrl.includes("firebasestorage")) {
+            console.log("Deleting:", imageUrl);
+            let imagePath = imageUrl.replace(baseUrl, "");
             const indexOfEndPath = imagePath.indexOf("?");
-            imagePath = imagePath.substring(0,indexOfEndPath);
-            imagePath = imagePath.replace(/%2F/g,"/");
-            imagePath = imagePath.replace(/%20/g," ");
+            imagePath = imagePath.substring(0, indexOfEndPath);
+            imagePath = imagePath.replace(/%2F/g, "/");
+            imagePath = imagePath.replace(/%20/g, " ");
             deleteObject(fsref(storage, imagePath)).catch((error) => {
-                console.log(error)
-            })
+              console.log(error);
+            });
           }
-      })
+        });
+      }
     },
   },
 };
